Created cleaner file paths ORGANIZED!
